+++ DigitizingDataAdminApp/Models/VslaMeetingInformation.cs
@@ -13,21 +13,21 @@ namespace DigitizingDataAdminApp.Models
     public class VslaMeetingInformation
     {
         public int MeetingId { get; set; }
-        public decimal cashExpenses { get; set; }
-        public decimal cashFromBank { get; set; }
-        public decimal cashFromBox { get; set; }
-        public decimal cashSavedBank { get; set; }
-        public decimal cashSavedBox { get; set; }
+        public long cashExpenses { get; set; }
+        public long cashFromBank { get; set; }
+        public long cashFromBox { get; set; }
+        public long cashSavedBank { get; set; }
+        public long cashSavedBox { get; set; }
         public decimal cashWelfare { get; set; }
         public DateTime? dateSent { get; set; }
         public Boolean isCurrent { get; set; }
         public Boolean isDataSent { get; set; }
         public int cycleId { get; set; }
-        public decimal cashFines { get; set; }
+        public long cashFines { get; set; }
         public DateTime? meetingDate { get; set; }
         public int membersPresent { get; set; }
-        public decimal totalSavings { get; set; }
-        public decimal totalLoans { get; set; }
-        public decimal totalLoanRepayment { get; set; }
+        public long totalSavings { get; set; }
+        public long totalLoans { get; set; }
+        public long totalLoanRepayment { get; set; }
     }
 }
